Predicted armor (#20560)
