+++ Content.Client/Armor/ArmorSystem.cs
@@ -1,0 +1,9 @@
+﻿using Content.Shared.Armor;
+
+namespace Content.Client.Armor;
+
+/// <inheritdoc/>
+public sealed class ArmorSystem : SharedArmorSystem
+{
+
+}
