+++ Resources/Prototypes/Entities/Objects/Vehicles/buckleable.yml
@@ -39,7 +39,7 @@
         sound:
           path: /Audio/Effects/metalbreak.ogg
       - !type:ExplodeBehavior
-  
+
 - type: entity
   parent: BaseVehicle
   id: BaseVehicleRideable
@@ -198,7 +198,7 @@
     baseSprintSpeed: 6
   - type: Armor
     modifiers:
-      coeffecients:
+      coefficients:
         Blunt: 0.8
         Slash: 0.6
         Piercing: 0.85
@@ -289,7 +289,7 @@
       maxBuckleDistance: 1
     - type: Armor
       modifiers:
-        coeffecients:
+        coefficients:
           Blunt: 0.8
           Slash: 0.6
           Piercing: 0.85
